Move taxon browsing helper methods into a separate file
